% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DHSHarmonization-package.R
\docType{package}
\name{DHSHarmonization-package}
\alias{DHSHarmonization}
\alias{DHSHarmonization-package}
\title{DHSHarmonization: DHS Harmonization Pipeline}
\description{
This package documents the pipeline used to merge and harmonize the DHS data for the lab.
}
\author{
\strong{Maintainer}: Tinashe Tapera \email{ttapera@hsph.harvard.edu}

Authors:
\itemize{
  \item Tinashe Tapera \email{ttapera@hsph.harvard.edu}
}

Other contributors:
\itemize{
  \item Christopher Golden \email{golden@hsph.harvard.edu} [copyright holder]
}

}
\keyword{internal}
