% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_raw_flat_dhs.R
\name{list_raw_flat_dhs}
\alias{list_raw_flat_dhs}
\title{Title}
\usage{
list_raw_flat_dhs(file_list)
}
\arguments{
\item{file_list}{Character vector of file paths to check.}
}
\value{
Character vector of flat DHS recode file paths.
}
\description{
Ingest the list of files in the project and return those that are flat DHS recode files
}
\examples{
link_inputs("config.yml") \%>\%
  list_raw_flat_dhs()
}
