% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_raw_gps_covars.R
\name{list_raw_gps_covars}
\alias{list_raw_gps_covars}
\title{Title}
\usage{
list_raw_gps_covars(file_list)
}
\description{
Description
}
\examples{
link_inputs("config.yml") \%>\%
  list_raw_gps_covars()
}
