% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_flat_dhs_data.R
\name{load_flat_dhs_data}
\alias{load_flat_dhs_data}
\title{load_flat_dhs_data Title}
\usage{
load_flat_dhs_data(folder_path)
}
\value{
1
}
\description{
load_flat_dhs_data Title
}
\examples{
ex_data <- here("data", "DHS Data", "DHS 1992", "MDIR21FL")
load_flat_dhs_data(ex_data) \%>\%
  tibble() \%>\%
  head()
}
