% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_gps_covars.R
\name{load_gps_covars}
\alias{load_gps_covars}
\title{Title}
\usage{
load_gps_covars(
  fpath,
  factor_vars = c("DHSCLUST", "DHSID", "DHSCC", "GPS_Dataset")
)
}
\description{
Description
}
\examples{
ex_data <- here("data", "DHS Data", "DHS 1997", "GPS Data", "MDGC32FL") \%>\%
  list.files(full.names = TRUE) \%>\%
  grep(pattern = "csv$", value = TRUE)
load_gps_covars(ex_data) \%>\%
  skimr::skim()
}
