% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_targets_pipeline.R
\name{build_targets_pipeline}
\alias{build_targets_pipeline}
\title{build_targets_pipeline Title}
\usage{
build_targets_pipeline(
  targets_notebook_name = "_target_notebook.Rmd",
  targets_file_name = "_targets.R",
  pkg_name = "DHSHarmonization",
  vignette_name = "pipeline-definition"
)
}
\value{
1
}
\description{
Dynamically build the targets pipeline for the project.
}
\details{
This function sets up the targets pipeline by defining a local \verb{_targets.R} file
and using \code{tar_tangle()} to create the necessary targets from an Rmd file.
The Rmd file is, itself, this vignette — Pipeline Definition.
}
\examples{
#build_targets_pipeline()
}
